+++ NEWS.md
@@ -1,7 +1,16 @@
 # cjbart 0.3.0
 
-* Add option to drop attributes from IMCE analysis
-* Add new pIMCE() function to calculate population-weighted IMCEs
+* Rename arguments in `het_vimp()` from `model` to `imces` and `outcomes` to `levels` for consistency across functions
+* Change `het_vimp()` so it returns a list, including the results and an attribute-level look-up table
+* Added ability to run `het_vimp()` calculation in parallel
+* Fixed bug when predicting continuous conjoint outcomes
+* Add new `pIMCE()` function to calculate population individual-level marginal component effects
+* Add new `AMCE()` function that includes Bayesian uncertainty estimate
+* Add check for unique levels across attributes. If not, the attribute-level values are modified to ensure uniqueness
+* Add functionality to improve quality of outputs
+* Update the vignette to reflect new functionality
+* Add ability to drop attributes from IMCE analysis
+* Remove dependencies not needed
 
 # cjbart 0.2.2